pyproject.toml
[tool.poetry]
name = "detectpii"
version = "0.1.0"
description = ""
authors = ["Fasih Khatib <hellofasih.confound928@passinbox.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
attrs = "^23.2.0"
cattrs = "^23.2.3"
trino = {extras = ["sqlalchemy"], version = "^0.329.0"}
psycopg2-binary = "^2.9.9"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
